web/requirements.txt
Flask==1.1.2
flask-redis==0.4.0
mysql-connector-python==8.0.25
mysql-connector-python-rf==2.2.2
urllib3==1.25.8
requests==2.25.1
beautifulsoup4==4.9.3
selenium==3.141.0
lxml==4.6.3